nodes: move children apply function to scene
This is only used here now.
