+++ libnopegl/src/scene.c
@@ -24,9 +24,63 @@
 #include "internal.h"
 #include "log.h"
 #include "memory.h"
+#include "params.h"
 #include "utils.h"
 
 NGLI_RC_CHECK_STRUCT(ngl_scene);
+
+typedef int (*children_func_type)(void *user_arg, struct ngl_node *parent, struct ngl_node *node);
+
+/*
+ * Apply a function on all children by walking through them. This is useful when
+ * node->children is not yet initialized (or confirmed to be complete yet).
+ */
+static int children_apply_func(children_func_type func, void *user_arg, struct ngl_node *node)
+{
+    uint8_t *base_ptr = node->opts;
+    const struct node_param *par = node->cls->params;
+
+    if (!par)
+        return 0;
+
+    while (par->key) {
+        uint8_t *parp = base_ptr + par->offset;
+
+        if (par->type == NGLI_PARAM_TYPE_NODE || (par->flags & NGLI_PARAM_FLAG_ALLOW_NODE)) {
+            struct ngl_node *child = *(struct ngl_node **)parp;
+            if (child) {
+                int ret = func(user_arg, node, child);
+                if (ret < 0)
+                    return ret;
+            }
+        } else if (par->type == NGLI_PARAM_TYPE_NODELIST) {
+            uint8_t *elems_p = parp;
+            uint8_t *nb_elems_p = parp + sizeof(struct ngl_node **);
+            struct ngl_node **elems = *(struct ngl_node ***)elems_p;
+            const size_t nb_elems = *(size_t *)nb_elems_p;
+            for (size_t i = 0; i < nb_elems; i++) {
+                struct ngl_node *child = elems[i];
+                int ret = func(user_arg, node, child);
+                if (ret < 0)
+                    return ret;
+            }
+        } else if (par->type == NGLI_PARAM_TYPE_NODEDICT) {
+            struct hmap *hmap = *(struct hmap **)parp;
+            if (hmap) {
+                const struct hmap_entry *entry = NULL;
+                while ((entry = ngli_hmap_next(hmap, entry))) {
+                    struct ngl_node *child = entry->data;
+                    int ret = func(user_arg, node, child);
+                    if (ret < 0)
+                        return ret;
+                }
+            }
+        }
+        par++;
+    }
+
+    return 0;
+}
 
 static int reset_nodes(void *user_arg, struct ngl_node *parent, struct ngl_node *node)
 {
@@ -45,7 +99,7 @@ static int reset_nodes(void *user_arg, struct ngl_node *parent, struct ngl_node 
 
     ngli_assert(!node->ctx);
 
-    int ret = ngli_node_children_apply_func(reset_nodes, s, node);
+    int ret = children_apply_func(reset_nodes, s, node);
     ngli_assert(ret == 0);
 
     ngli_darray_reset(&node->children);
@@ -81,7 +135,7 @@ static int setup_nodes(void *user_arg, struct ngl_node *parent, struct ngl_node 
         ngli_darray_init(&node->children, sizeof(struct ngl_node *), 0);
         ngli_darray_init(&node->parents, sizeof(struct ngl_node *), 0);
 
-        int ret = ngli_node_children_apply_func(setup_nodes, s, node);
+        int ret = children_apply_func(setup_nodes, s, node);
         if (ret < 0)
             return ret;
     }
@@ -218,7 +272,7 @@ static int find_livectls(void *user_arg, struct ngl_node *parent, struct ngl_nod
         }
     }
 
-    return ngli_node_children_apply_func(find_livectls, hm, node);
+    return children_apply_func(find_livectls, hm, node);
 }
 
 int ngl_livectls_get(struct ngl_scene *scene, size_t *nb_livectlsp, struct ngl_livectl **livectlsp)
